Removing the need for a key in the trigger descriptor, and as such, removing the trigger descriptor class.
I called the fact that there was way too much abstraction happening. After not looking at this code for about a good 3 or 4 months I couldn't really understand why I needed such a lame class. That, and the fact that I was basically forcing my trigger string to have an unneded "key" really made me mad.
So, I removed the key, which takes away the need to even parse that part of the string. Now the trigger descriptor is just a single value that doesn't need to be handled by any resource other than the trigger, which is where it should be.