Removing the general test-user-awards-listener.php file and separating out the test logic.
The three types of listener functionality we were testing were on whether or not the listener could give awards based on whether user meta was assigned, created, or updated on a user.
We have separated this out into three specific files that allow us to add tests that more specifically go towards this functionality.